Optimizations

+++ include/equalizer.h
@@ -10,9 +10,9 @@
 
 #include "filter_def.h"
 
-uint8_t eq_check_f0_variation(filter_instance* S, float32_t f0);
+uint8_t eq_check_f0_variation(filter_instance* S, uint16_t f0);
 uint8_t eq_check_Q_variation(filter_instance* S, float32_t Q);
-uint8_t eq_check_G_variation(filter_instance* S, float32_t G);
+uint8_t eq_check_G_variation(filter_instance* S, int8_t G);
 uint8_t eq_coef_calc(filter_instance *SL);
 
 #endif /* EQUALIZER_H_ */
